mlops_pbi/pyproject.toml
[tool.poetry]
name = "mlops_pbi"
version = "0.1.0"
description = ""
authors = ["Sanmi Ibitoye"]

[tool.poetry.dependencies]
python = "^3.8"
dvc = "^2.10.0"
hydra-core = "^1.1.1"
pdoc3 = "^0.10.0"
notebook = "^6.4.10"
pre-commit = "^2.19.0"
pandas = "^1.5.2"
prefect = "^2.7.0"
jupyter = "^1.0.0"
seaborn = "^0.12.1"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build   
)/ 
'''
